// generated/src/app/api-models/auth-second-opinion.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthSecondOpinion {
  authNumber: number;
  secondaryAuthNo: string;
  seqSecondOp: number;
  seqProvId: number;
  secondOpinionDate: Date;
  secondOpinion: string;
  decision: string;
  comments: string;
  status: string;
  statusReason: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}